Backport 32723b8 from pika
 Do not disconnect on SSL_ERROR_WANT_WRITE

In certain cases, when using Python 3.5+, `SSLSocket.send` may
raise `ssl.SSL_ERROR_WANT_WRITE` when the socket is not currently
able to write. The previous behavior would close the connection.
The new behavior is to treat this just like an `EWOULDBLOCK`
socket error which is what this is analogous to.